Make ActionHandler compatible with generic routers.
Like those found at https://github.com/middlewares/awesome-psr15-middlewares#router. This essentially converts this middleware package into an ADR-specific alternative to https://github.com/middlewares/request-handler.
